Add files via upload
Waggle company project, a startup that makes smart devices for pets.
I  have been tasked with delivering a boardroom-ready Power BI report that tells the story of how the Lapcat data compares to findings from the dog collar Lapdog devices.
Business Requests:

The CEO is curious about the following questions:
Did the average daily steps increase for cats wearing the device as they did for dogs?
Were owners of Lapcat devices as satisfied with the product as Lapdog owners?
The Chief Marketing Officer would like your report to be “on-brand” by including only colors from the Waggle color palette, the Waggle logo, and other approved company logos and icons.
The product team wants to incorporate other visuals and insights  but is most interested in demographic comparisons between the dogs and cats using Waggle devices as well as any information about the families who own the pets. there are slicers to help the team filter and explore on their own.